fix: msvc warnings (#4651)
* fix: C4189 local variable is initialized but not referenced

* fix: C4706 assignment within conditional expression

* fix: C4018 signed/unsigned mismatch

* fix: warning C4996 High-DPI scaling is always enabled.

* fix: C4996: QApplication::fontMetrics() is deprecated

https://github.com/qt/qtbase/commit/eed5514eaf03c04f1b93d0899789841b8488d9a6

+++ tests/libtransmission/announcer-udp-test.cc
@@ -409,7 +409,7 @@ TEST_F(AnnouncerUdpTest, canMultiScrape)
     auto buf = libtransmission::Buffer{};
     buf.addUint32(ScrapeAction);
     buf.addUint32(scrape_transaction_id);
-    for (size_t i = 0; i < expected_response.row_count; ++i)
+    for (int i = 0; i < expected_response.row_count; ++i)
     {
         buf.addUint32(expected_response.rows[i].seeders);
         buf.addUint32(expected_response.rows[i].downloads);
